Some refactoring to ready match-boss-catalogue.py
